Extracting the traits from the downloaded data.

Extracting the traits from the downloaded data.

fw_extract(
  data,
  taxa,
  ecotraits = NULL,
  taxaorder = NULL,
  token = NULL,
  multiple = FALSE,
  parallel = FALSE,
  cores = NULL,
  quietly = FALSE,
  pct = 80,
  errorness = 20,
  warn = FALSE
)

Arguments

data

vector. The list or vector with species names for which ecological references needs to be extracted from the database.

taxa

string. The taxa group to download from the platform. The allowed group include "fi", "mi", "pp", "pb", "di","mp" for fishes, macroinvertebrates, phytoplankton, phytobenthos, diatoms, and macrophytes. Multiple groups allowed such as 'pp', 'di'.

pp: Pytoplankton.

mp: Macrophtytes

mi: Macroinvertebrates

fi: Fishes

di: Diatoms

pb: Phytobenthos without diatoms

ecotraits

vector. Selected traits that should be downloaded for particular taxa group. Check fw_ecoparamdb for the allowed traits in the database.

taxaorder

vector. If taxa is mi, the taxaorder must be indicated for data to be downloaded. The different macroinvertebrates orders allowed can be obtained using fw_orders function.

token

string. This is a required parameter to allow user authentication with the platform. To get the token, use before_u_start function to get the required steps. Remember that the token is saved in memory such that the data downloaded is not re-downloaded in the next session.

multiple

logical. Either TRUE if multiple taxa groups are considered. Default is FALSE for only single taxon group.

parallel

logcial. Either TRUE to allow parallelisation most especially if multiple groups are considered. If multiple is set to TRUE and parallel to FALSE, a slow mode of data download will be executed.

cores

interger. The number computer cores to be used to run while data download and its necessary only when parallel is set to TRUE. Default is 2.

quietly

logical. If FALSE, a message showing the slow mode of data collation will be presented to the user. Default TRUE.

pct

numeric. The number used as a cutoff to infer similarity of the user provided name and what is found in the database. The higher the percentage, the higher the similarity the species name provided by the user and the one in the database. pct ranges from 0 to 100 but the default is 80 to ensure that wrong names are not selected at low similarity percentage cutoff.

errorness

numeric Similar to pct, errorness parameter uses the distance differences between the user-provided names and all the taxa group species standard names. The lower the percentage error, the higher the similarity in the species names provided. Default is 20 and beyond 30, a warning is showed to avoid wrong species replace the user provided name, which leads to extracting wrong traits.

warn

logical To show species name warning checks and traits cleaning. Default FALSE.

Value

dataframe A dataframe species traits for all orders.