Data download from the Freshwaterecology.info database.

Data download from the Freshwaterecology.info database.

fw_pulldata(
  taxa,
  ecotraits = NULL,
  taxaorder = NULL,
  token = NULL,
  warn = FALSE,
  multiple = FALSE,
  parallel = FALSE,
  cores = 2,
  quietly = FALSE
)

Arguments

taxa

string. The taxa group to download from the platform. The allowed group include "fi", "mi", "pp", "pb", "di","mp" for fishes, macroinvertebrates, phytoplankton, phytobenthos, diatoms, and macrophytes. Multiple groups allowed such as 'pp', 'di'.

pp: Pytoplankton.

mp: Macrophtytes

mi: Macroinvertebrates

fi: Fishes

di: Diatoms

pb: Phytobenthos without diatoms

ecotraits

vector. Selected traits that should be downloaded for particular taxa group. Check fw_ecoparamdb for the allowed traits in the database.

taxaorder

vector. If taxa is mi, the taxaorder must be indicated for data to be downloaded. The different macroinvertebrates orders allowed can be obtained using fw_orders function.

token

string. This is a required parameter to allow user authentication with the platform. To get the token, use before_u_start function to get the required steps. Remember that the token is saved in memory such that the data downloaded is not re-downloaded in the next session.

warn

logical To show species name warning checks and traits cleaning. Default FALSE.

multiple

logical. Either TRUE if multiple taxa groups are considered. Default is FALSE for only single taxon group.

parallel

logcial. Either TRUE to allow parallelisation most especially if multiple groups are considered. If multiple is set to TRUE and parallel to FALSE, a slow mode of data download will be executed.

cores

interger. The number computer cores to be used to run while data download and its necessary only when parallel is set to TRUE. Default is 2.

quietly

logical. If FALSE, a message showing the slow mode of data collation will be presented to the user. Default TRUE.

Value

list. Lists of taxa groups data downloaded from the Freshwater Information Platform.

Details

The getdata is used as a wrapper to enable the user to download data from multiple groups and allowing parallelising the getfiles function. The user can change the number of cores to improve the speed of data downloaded. All data downloaded in this function by the user is cached in the taxongroups folder, which is automatically created in the working directory when loading the package. The data is permanently cached, even when the session is restarted. The user can use the fw_decache function to remove the files. However, to extract the data please check both extract_traits and fw_ecoparameters.

Examples

if (FALSE) { # \dontrun{

ppdata <- 1
} # }