To download data from the Freshwaterecology.info database.

The function provides a seamless access and download of species ecological parameters, traits, or indicators from the Freshwaterecology.info database. The function allows multiple inclusion of organism groups, which include macroinvertebrates, fishes, phytoplankton, phytobenthos, macrophytes, and diatoms.

Usage

fw_searchdata(
  organismgroup,
  taxa_searched = NULL,
  ecoparams = NULL,
  apikey = NULL,
  warn = TRUE,
  seed = 1135,
  secure = TRUE,
  inform = FALSE,
  taxalevel = NULL,
  cachefolder = "cache"
)

Arguments

organismgroup

string. The organism group to download from the platform. The allowed group include "fi", "mi", "pp", "pb", "di","mp" for fishes, macroinvertebrates, phytoplankton, phytobenthos, diatoms, and macrophytes, respectively. Multiple groups allowed such as 'pp', 'di'.

pp: Pytoplankton.

mp: Macrophtytes

mi: Macroinvertebrates

fi: Fishes

di: Diatoms

pb: Phytobenthos without diatoms

taxa_searched

string An internal placeholder to accommodate the standard taxonomic names for invertebrates and phytobenthos from the database.

ecoparams

vector. Selected traits that should be downloaded for particular organism group. Check fw_dbguide for the allowed traits in the database.

apikey

string. The API key which is automatically loaded using the loadapikey() internal function.

warn

logical To show species name warning checks and traits cleaning. Default FALSE.

seed

integer. An integer to help track the caching of the access token generated during data collation. If a user wants to get a new token, then the seed should be changed.

secure

logical. If TRUE, the user will be prompted to set the API key in the .Renviron file by running the fw_setapikey function. The User must strictly type in API_KEY = 'api key', save, close the file and restart the R session or RStudio for the API_KEY environment to be captured. If FALSE, then the key will be entered directly in the API_KEY directly in the fw_token() function. This method is insecure, since the key can be obtained from the codes by other users.

inform

logical. To indicate if the token is successfully generated. Default TRUE.

taxalevel

string Allowed taxonomic levels at which data can retrieved. Default is 'species' but data can also be downloaded at family level, genus, and taxa group level.

cachefolder

string. The root path were the cached data will be saved on the user PC. If the path is not provided, the cached information will be saved in the current working directly.

Value

List of download species traits

Details

For macroinvertebrates, since there is alot of data which take a lot of time to download and yet slowing the process, the taxa_searched parameter should be provided to only search for the particular traits, family and orders or taxa group where the species falls. Also, the phytobenthos requires to provide the taxa_searched to enable search for a particular taxa group. For other organism group, the parameter should not provided because will not be used in the data search.

Examples

if (FALSE) { # \dontrun{
dfsearch <- fw_searchdata(organismgroup = 'fi', ecoparams = 'migration', cachefolder = 'cache')
} # }